backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
pandas==2.2.2
google-generativeai==0.8.1
twilio==9.3.0
apscheduler==3.10.4
aiofiles==24.1.0
python-dotenv==1.0.1
httpx==0.27.2
structlog==24.4.0
tenacity==9.0.0
aiosqlite==0.20.0
pytz==2024.2
